added ringcentral_ to all our functions and ...
updated the userguide, tested the STOP SMS option, the email and SMS opt-in changes as well. The re-activation code for the plugin to create / update tables, updated the un-install process to remove the new DB table, increased the version # to 1.3.0, fixed a few type-os in the readme file, general code clean up.